test(ocapn): move test failure error unwrapping to test utils
